Add PID group with pitch, roll and yaw
The PIDGroup is generic and three instances of it are created within
the main group. Each PIDGroup is pushed against the left side of the
main group.

Each PIDGroup consists of a title and a matrix of values. It also
initializes an array for storing the value of the control (like
pitch), and kp, ki and kd constants.