Implement parse function for input read from the Arduino
The parse function was using a regular file up until now. This commit
updates the parse function to use the SerialPort gem to read from the
Arduino.

The output of the Arduino is of the form:

^Xa Ya Za Xd Yd Zd Throttle Pkp Pki Pkd Rkp Rki Rkd Ykp Yki Ykd m0 m1 m2 m3$
--------- --------          ----------- ----------- ----------- -----------
actual   desired            Pitch PID   Roll PID     Yaw PID   Motor Setting

The values are space separated.
^ signifies the beginning of the line
$ signifies the end of the line.

The PID values give the current PID constants set for each control
axis.  The motor settings represent the actual value that each motor
is set to in the following order:

 Front (Red hands of the frame)
m3    m0
  \  /
   \/
   /\
  /  \
m2    m1

The values range from 0 to 180.

The desired angles, throttle, PID settings are readbacks that were
sent to the Arduino.