Update PIDGroup to write out and refactor
When the PIDGroup variables are updated, the corresponding command and
value are sent out to the Arduino.

The Text Field has been replaced with a Real Spinner, ie., arrows that
control that give easy increment and decrement capabilities.
The granularity is 0.01.

The code has been refactored to remove unnecessary repetition.

The format of the commands is:
ck
where c can be 'p', 'r', or 'y' for pitch, roll and yaw, respectively.
and k can be 'p', 'i', or 'd' for proportional, integral and derivative.