Remove motor gauges
We do not expect any information back from the quadcopter. Without
this telemetry, showing the motor gauges are not possible.

Another issue is that this feedback is absolutely no use to the pilot
since he cannot monitor these and fly the quadcopter at the same time.

It might be useful instead to include a throttle bar, and perhaps
an orientation feedback for debugging.