Move FlightControl variables to DashboardWindow
FlightControl variables (FXDataTargets) are stored in the
FlightControlGroup objects. Ideally, the GUI element should be
separated from the actual variable so that it is easier for the
variable to be updated from multiple locations.

Move the FXDataTargets to the DashboardWindow object to allow the
variable's modification using other methods as well. (This is in
expectation of the addition of throttle control using the mousewheel.)