Add Mousewheel control for throttle
In a move to use the dashboard to control the quadcopter completely,
the requirement to change parameters such as throttle arises.

In this implementation, the scroll wheel on the mouse (or scrolling
on a touchpad) can be used to control the throttle of the quadcopter.

Since mouse events from anywhere on the window cannot be captured
easily, an explicit canvas is created *over which the mouse must
be present* for scroll control to work.

There is also a provision for a kill switch that immediately sets
the throttle to 0 when the spacebar is pressed.

The canvas must be active to receive the spacebar press event. This
is achieved by clicking on it once initially.

Warning:
1) The canvas must be selected before key events (spacebar) are sent
   to it.
2) The mousewheel scroll event is only sent to the canvas when the
   mouse is over it.